// src/constants/api.d.ts
export declare const BASE_URL = "https://abc.wesap.vn";
export declare const SOCKET_URL = "wss://abc.wesap.vn";
export declare const AUTH_URL: {
    REGISTER_URL: string;
    LOGIN_URL: string;
    REFRESH_URL: string;
    VERIFY_OTP_URL: string;
    RESENT_OTP_URL: string;
    LOGIN_GOOGLE_URL: string;
    SEND_OTP_FOR_RESET_PASS_URL: string;
    RESET_PASSWORD_URL: string;
    CHANGE_PASSWORD_URL: string;
};
export declare const PROFILE_URL: {
    UPDATE_PROFILE_URL: string;
    GET_PROFILE_URL: string;
    GET_OTHER_PROFILE_URL: string;
};
export declare const MEDIA_URL: {
    UPDATE_MEDIA_URL: string;
};
export declare const POST_URL: {
    CREATE_POST_URL: string;
    GET_MY_POSTS_URL: string;
    GET_POST_URL: string;
    LIKE_POST_URL: string;
    GET_COMMENTS_BY_POST_ID_URL: string;
    GET_REPLY_COMMENTS_URL: string;
    REPLY_COMMENTS_URL: string;
    COMMENT_POST_URL: string;
    SHARE_POST_URL: string;
    ARCHIVE_POST_URL: string;
    DELETE_POST_URL: string;
    EDIT_POST_URL: string;
    GET_MY_SHARE_POST_URL: string;
    GET_OTHER_SHARE_POST_URL: string;
    GET_OTHER_POST_URL: string;
};
export declare const CHAT_URL: {
    CHAT_URL: string;
    CHAT_MESSAGE_URL: string;
};
export declare const NOTIFICATION_URL: {
    GET_ALL_NOTIFICATIONS_URL: string;
    MARK_AS_READ_URL: string;
};
export declare const FRIEND_URL: {
    GET_FRIEND_REQUEST: string;
    GET_LIST_FIENDS_URL: string;
    ACCEPT_FRIEND_URL: string;
    REJECT_FRIEND_URL: string;
    UNFRIEND_URL: string;
    SEND_FRIEND_REQUEST: string;
    SEARCH_USER_URL: string;
};